pyproject.toml
[project]
name = "AI-Portfolio-Chatbot"
version = "0.1.0"
description = "Un assistant IA conversationnel intelligent pour votre site web ou portfolio personnel."
requires-python = "==3.12.*"
dependencies = [
    "gradio>=5.49.1",
    "openai>=1.50.0",
    "openai-agents>=0.13.0",
    "pypdf>=5.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.30.0",
]
